Create deepseek
Installing and Deploying DeepSeek
DeepSeek is one of the most popular topics in this period. Its distillation model enables low-cost and high-quality inference and allows us to use foundation models on local miniaturized hardware. This exercise describes how to deploy the DeepSeek R1 mode
Objectives
This exercise is about the deployment and experience of DeepSeek 1.5B. The file size is about 1.1 GB, which is easy to download. In addition, when deployed and runs in the CPU environment, this version brings fluent performance for users.

In this exercise, you will learn how to use Ollama to deploy the DeepSeek R1 model.

Prerequisites
Use the exercise account to access Lab Desktop.

What is Lab Desktop?
Click the refresh button of the browser to refresh the exercise page.


On the exercise operation interface, open the browser, and go to the Huawei Cloud login page. Click IAM User. Then, use the account assigned to you to log in to Huawei Cloud.

Note: Use the account provided by the exercise tutorial.